Move NSView extension into its own file
